Make libopae-c-ase.so the duplicate of libopae-c.so in order to make … (#1269)
* Make libopae-c-ase.so the duplicate of libopae-c.so in order to make the ASE plugin changes backward compatible for ASE applications

* If WITH_ASE is set, it will fail loading the library when not finding the opae_ase.cfg file

* Set the environment variable WITH_ASE in init_ase.c
