+++ libopae/plugins/ase/scripts/with_ase
@@ -1,4 +1,4 @@
 #!/bin/sh
 
 # Run a binary using the ASE simulation library instead of the standard libopae-c.
-env LD_PRELOAD="libopae-c-ase.so${LD_PRELOAD:+:$LD_PRELOAD}" "$@"
+env WITH_ASE=1 "$@"
